WIP: Render atoms
The tricksy bit here is that we bottom out on children when we reach an
atom. Instead of rendering the value in the children array, we convert
it into a prop alongside the payload so they can be passed to the
component. This is a deviation from the established tree structure, but
it means atom components can take a keyed `value` prop instead of
needing to inspect `children`. I'd prefer to establish that children
shouldn't be accessed directly, as it maintains the pattern of
traversing and operating on each node, independently, in sequence.