Move React-specific logic to ReactRenderer, leave Tree generic
By returning an atom name instead of the actual atom component, we're
able to decouple Tree from React and leave it fairly generic. The React
renderer should handle mapping atom names to components. This way we do
as much as possible before handing off to the last-mile renderer, but
stop short of tying Tree to any specific implementation.